requirements.txt
Flask==3.1.0
torchvision==0.21.0
Pillow==11.1.0
werkzeug==3.1.3